+++ .github/scripts/cover2cover.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-#  Copyright Amazon.com Inc. or its affiliates.
+#  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #  SPDX-License-Identifier: Apache-2.0
 
 import sys
